nmap-automator/pyproject.toml
[tool.poetry]
name = "nmap-automator"
version = "0.1.3"
description = ""
authors = ["Elias Dargham <elias.dargham@net.usj.edu.lb>", "Anthony Tabet <anthony.tabet2@net.usj.edu.lb>"]
readme = "README.md"
packages = [{include = "nmap_automator", from = "src"}]

[tool.poetry.scripts]
nmap-automator = "nmap_automator.runner:main"

[tool.poetry.dependencies]
python = "^3.11"
python-nmap = "^0.7.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openai = "^1.58.1"
google-generativeai = "^0.8.3"
ollama = "^0.4.4"
pydantic = "^2.10.4"
omegaconf = "^2.3.0"
flask = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
